self code review, boudn.py
