+++ packages/nemo_platform_plugin/src/nemo_platform_plugin/client/types.py
@@ -12,7 +12,7 @@ from __future__ import annotations
 
 from collections.abc import AsyncIterable, Iterable
 from dataclasses import dataclass
-from typing import Generic, TypedDict, TypeVar
+from typing import Generic, Protocol, TypedDict, TypeVar, Unpack
 
 from pydantic import BaseModel
 
@@ -51,6 +51,34 @@ class BasePath(TypedDict):
 PathT = TypeVar("PathT", bound=BasePath)
 RequestT = TypeVar("RequestT", bound=BaseModel)
 ResponseT = TypeVar("ResponseT", bound=BaseModel | BinaryContent | Stream | None)
+
+
+# ---------------------------------------------------------------------------
+# Protocols — what bound callables need from endpoints
+# ---------------------------------------------------------------------------
+
+
+class BodyRequestable(Protocol[PathT, RequestT, ResponseT]):
+    """Protocol for endpoints that accept a JSON payload."""
+
+    def request(self, payload: RequestT, **path_params: Unpack[PathT]) -> PreparedRequest[ResponseT]: ...
+
+
+class BinaryBodyRequestable(Protocol[PathT, ResponseT]):
+    """Protocol for endpoints that accept binary content."""
+
+    def request(self, content: bytes | Iterable[bytes] | AsyncIterable[bytes], **path_params: Unpack[PathT]) -> PreparedRequest[ResponseT]: ...
+
+
+class NoBodyRequestable(Protocol[PathT, ResponseT]):
+    """Protocol for endpoints with no request body."""
+
+    def request(self, **path_params: Unpack[PathT]) -> PreparedRequest[ResponseT]: ...
+
+
+# ---------------------------------------------------------------------------
+# PreparedRequest
+# ---------------------------------------------------------------------------
 
 
 @dataclass(frozen=True, slots=True)
